requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0
PyJWT==2.8.0
Werkzeug==2.3.7

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
